double MySQL type for expense amount

+++ src-backend/db/create.sql
@@ -1,7 +1,7 @@
 create table expenses (
   `id` int not null auto_increment,
   `date` date not null,
-  `amount` decimal(4,2),
+  `amount` double,
   `reason` varchar(250) not null,
   primary key (`id`)
   );
